Files for exercises 3-4.
In exercise4 folder the simple-web-server contains a basic http server which gives different kinds of responses to clients depending the request path. exercise3/file-read-example is just an initial commit with an empty file. However this will be an example code which reads file in three different ways - synchronously, asynchronously with callback and asynchronously with a Promise.